fix diffusion_step warns

+++ src/multi_channel_mixed_feedback.rs
@@ -14,7 +14,7 @@ impl MultiChannelMixedFeedback {
         for c in 0..CHANNELS {
             let r = c as f64 * 1.0 / CHANNELS as f64;
             self.delaySamples[c] = (f64::powf(2., r) * delaySamplesBase) as i64;
-            self.delays[c].resize((self.delaySamples[c] + 1) as usize);
+            self.delays[c].resize(self.delaySamples[c] + 1);
             self.delays[c].reset();
         }
     }
